+++ contracts/multisig-prover/src/encoding/mod.rs
@@ -1,4 +1,5 @@
 // Adding more encoding modules will require using feature flags in here to switch between them
-pub mod evm;
+pub mod encoding;
+pub mod abi;
 
-pub use evm::Data;
+pub use encoding::Data;
